Convert AuthContext and SearchContext from typescript to javascript
